+++ splendor_env/agents/__init__.py
@@ -2,5 +2,6 @@
 
 from .greedy_agent import GreedyAgent
 from .random_agent import RandomLegalAgent
+from .shortest_agent import ShortestAgent
 
-__all__ = ["GreedyAgent", "RandomLegalAgent"]
+__all__ = ["GreedyAgent", "RandomLegalAgent", "ShortestAgent"]